+++ mt_gui/src/subPage1/OrderDetailPanel.vue
